Consolidate `spine-protoc-plugin` location in one class
