+++ tools/plugin-base/src/main/java/io/spine/tools/gradle/ProtocConfigurationPlugin.java
@@ -21,7 +21,6 @@
 package io.spine.tools.gradle;
 
 import com.google.common.collect.ImmutableList;
-import com.google.errorprone.annotations.OverridingMethodsMustInvokeSuper;
 import com.google.protobuf.gradle.ExecutableLocator;
 import com.google.protobuf.gradle.GenerateProtoTask;
 import com.google.protobuf.gradle.ProtobufConfigurator;
@@ -37,9 +36,7 @@ import java.util.Collection;
 
 import static io.spine.tools.gradle.ProtobufDependencies.gradlePlugin;
 import static io.spine.tools.gradle.ProtobufDependencies.protobufCompiler;
-import static io.spine.tools.gradle.ProtocPluginName.spineProtoc;
 import static io.spine.tools.groovy.ConsumerClosure.closure;
-import static org.gradle.internal.os.OperatingSystem.current;
 
 /**
  * An abstract base for Gradle plugins that configure Protobuf compilation.
@@ -48,11 +45,6 @@ import static org.gradle.internal.os.OperatingSystem.current;
  * no action is performed.
  */
 public abstract class ProtocConfigurationPlugin extends SpinePlugin {
-
-    protected static final String SPINE_PLUGIN_NAME = "spine-protoc-plugin";
-    private static final String SH_EXTENSION = "sh";
-    private static final String BAT_EXTENSION = "bat";
-    private static final String SCRIPT_CLASSIFIER = "script";
 
     protected static final DependencyVersions VERSIONS = DependencyVersions.get();
 
@@ -107,28 +99,8 @@ public abstract class ProtocConfigurationPlugin extends SpinePlugin {
      *         which
      *         is a required plugin
      */
-    @OverridingMethodsMustInvokeSuper
-    protected void configureProtocPlugins(NamedDomainObjectContainer<ExecutableLocator> plugins) {
-        if (!spineProtocIsPresent(plugins)) {
-            plugins.create(spineProtoc.name(), locator -> {
-                boolean windows = current().isWindows();
-                String scriptExt = windows ? BAT_EXTENSION : SH_EXTENSION;
-                locator.setArtifact(Artifact.newBuilder()
-                                            .useSpineToolsGroup()
-                                            .setName(SPINE_PLUGIN_NAME)
-                                            .setVersion(VERSIONS.spineBase())
-                                            .setClassifier(SCRIPT_CLASSIFIER)
-                                            .setExtension(scriptExt)
-                                            .build()
-                                            .notation());
-            });
-        }
-    }
-
-    private static boolean
-    spineProtocIsPresent(NamedDomainObjectContainer<ExecutableLocator> plugins) {
-        return plugins.findByName(spineProtoc.name()) != null;
-    }
+    protected abstract void
+    configureProtocPlugins(NamedDomainObjectContainer<ExecutableLocator> plugins);
 
     private void configureProtocTask(GenerateProtoTask protocTask) {
         configureDescriptorSetGeneration(protocTask);
